No use having the Processor base class not in the processor.py module.
