+++ pypond/sources.py
@@ -127,42 +127,3 @@ def add_prev_to_chain(n, chain):  # pylint: disable=invalid-name
     else:
         add_prev_to_chain(n.prev(), chain)
 
-
-class Processor(Observable):
-    """
-    Base class for all pipeline processors.
-    """
-
-    def __init__(self, arg1, options):
-        super(Processor, self).__init__()
-
-        self._log('Processor.init')
-
-        self._id = unique_id('processor-')
-
-        self._pipeline = None
-        self._prev = None
-
-        if is_pipeline(arg1):
-            self._pipeline = arg1
-            self._prev = options.prev
-
-    def prev(self):
-        """Return prev"""
-        return self._prev
-
-    def pipeline(self):
-        """Return the pipeline"""
-        return self._pipeline
-
-    def chain(self):
-        """Return the chain"""
-        chain = [self]
-
-        if is_pipeline(self.prev()):
-            chain.append(self.prev().input())
-            return chain
-        else:
-            return add_prev_to_chain(self.prev(), chain)
-
-    # flush() is inherited from Observable
